add sprite.move() / onUpdate(timePerFrame) make timePreFrame an 'Integer'

+++ bow/bow-engine-demo/src/main/java/basics/components/BulletComponent.java
@@ -27,7 +27,7 @@ public class BulletComponent extends CloneableComponent {
     }
 
     @Override
-    public void onUpdate(double timePerFrame) {
+    public void onUpdate(int timePerFrame) {
 
     }
 }
