+++ bow/bow-engine/src/main/java/com/pokewords/framework/sprites/components/CollidableComponent.java
@@ -22,7 +22,7 @@ public class CollidableComponent extends Component implements Shareable {
 	}
 
 	@Override
-	public void onUpdate(double timePerFrame) { }
+	public void onUpdate(int timePerFrame) { }
 
 	@Override
 	public void onAppStateEnter() { }
