+++ bow/bow-engine/src/main/java/com/pokewords/framework/sprites/components/GameLifecycleListener.java
@@ -6,7 +6,7 @@ public interface GameLifecycleListener {
      * triggered in every game-loop
      * @param timePerFrame time per frame
      */
-    void onUpdate(double timePerFrame);
+    void onUpdate(int timePerFrame);
