+++ bow/bow-engine/src/main/java/com/pokewords/framework/sprites/components/StringsComponent.java
@@ -31,7 +31,7 @@ public class StringsComponent extends Component implements Shareable, Renderable
     public void onAppStateDestroy() { }
 
     @Override
-    public void onUpdate(double timePerFrame) { }
+    public void onUpdate(int timePerFrame) { }
 
     @Override
     public List<StringFrame> getRenderedFrames() {
